Initial Commit -
Initial commit for Brick Breaker game for Assignment 3 -

Spent time familiarising myself with the overall structure of the
program. Noticed a number of similarities with previous programs

Have added a few comments onto the SETTINGS class - these are for
description only and offer no functional purpose

Next Task:

Is to set the BALL and PADDLE classes up, that includes setting their
physical properties, initial start positions etc.

Will also work on the SPRITE class which houses a number of shared
properties for both the BALL and PADDLE classes.

All necessary variables are being housed in the SETTINGS class - are set
up as FINAL so will not be edited any further.