Third Commit -
nearly all coding is finished.. only code left to write is for the game
won or lost messages in the BreakoutPanel class...

Game is not working as expected however - bricks are only element being
painted onto the panel and even then, they are being painted in the top
left hand corner stacked on top of each other.

unsure if paddle or ball element are being painted as well.. probably
not as those elements have vectors and neither moves when on screen.

Need to go through code line by line to determine next step